requirements.txt
accelerate==0.22.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.13.1
altair==5.2.0
annotated-types==0.5.0
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.3.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.14.0
backcall==0.2.0
beautifulsoup4==4.12.2
bitsandbytes==0.41.1
bleach==6.1.0
blinker==1.7.0
cachetools==5.3.3
catboost==1.2.3
certifi==2020.6.20
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.2.0
click==7.1.2
colorama==0.4.6
colorlog==6.8.2
comm==0.1.4
contourpy==1.1.0
cycler==0.11.0
dataclasses-json==0.5.14
datasets==2.14.4
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.7
distlib==0.3.6
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.19.1
filelock==3.12.1
Flask==1.1.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.42
graphviz==0.20.3
greenlet==2.0.2
huggingface-hub==0.16.4
idna==2.10
ipykernel==6.25.1
ipython==8.15.0
isoduration==20.11.0
itsdangerous==1.1.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.3.1
jupyter_core==5.3.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.0.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.5
langchain==0.0.274
langsmith==0.0.26
llvmlite==0.42.0
loralib==0.1.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nbclient==0.7.4
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==3.1
notebook_shim==0.2.3
numba==0.59.1
numexpr==2.8.5
numpy==1.25.2
openpyxl==3.1.2
optuna==3.6.0
overrides==7.4.0
psycopg2-binary==2.9.9
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
peft==0.5.0
phik==0.12.4
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.5.3
plotly==5.18.0
prometheus-client==0.19.0
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
pure-eval==0.2.2
pyarrow==13.0.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
pydeck==0.8.1b0
Pygments==2.16.1
pynndescent==0.5.11
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.14.0
python-json-logger==2.0.7
regex==2023.8.8
rich==13.7.1
rpds-py==0.16.2
safetensors==0.3.3
scikit-learn==1.4.1.post1
scipy==1.11.2
Send2Trash==1.8.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
stack-data==0.6.2
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
tensorboardX==2.6.2.2
terminado==0.18.0
threadpoolctl==3.4.0
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.32.0
types-python-dateutil==2.8.19.20240106
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
umap-learn==0.5.5
uri-template==1.3.0
urllib3==1.25.10
virtualenv==20.23.0
voila==0.5.5
watchdog==4.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==1.0.1
xlrd==2.0.1
xxhash==3.3.0
yarl==1.9.2
